# docs/swagger.yaml
components:
  schemas:
    Body_convert_files_to_dataframe_api_v1_text_upload_dataset_post:
      properties:
        csv_file:
          format: binary
          title: Csv File
          type: string
      required:
      - csv_file
      title: Body_convert_files_to_dataframe_api_v1_text_upload_dataset_post
      type: object
    Body_convert_files_to_dataframe_api_v1_upload_dataset_post:
      properties:
        json_file:
          format: binary
          title: Json File
          type: string
        zip_file:
          format: binary
          title: Zip File
          type: string
      required:
      - json_file
      - zip_file
      title: Body_convert_files_to_dataframe_api_v1_upload_dataset_post
      type: object
    Body_get_text_api_v1_text_get_text_post:
      properties:
        music_file:
          description: "MP3 \u0430\u0443\u0434\u0438\u043E\u0444\u0430\u0439\u043B\
            \ \u0434\u043B\u044F \u0440\u0430\u0441\u043F\u043E\u0437\u043D\u0430\u0432\
            \u0430\u043D\u0438\u044F"
          format: binary
          title: Music File
          type: string
      required:
      - music_file
      title: Body_get_text_api_v1_text_get_text_post
      type: object
    Body_predict_api_v1_predict_post:
      properties:
        data:
          format: binary
          title: Data
          type: string
        model_name:
          title: Model Name
          type: string
      required:
      - model_name
      - data
      title: Body_predict_api_v1_predict_post
      type: object
    Body_predict_by_music_file_api_v1_predict_mp3_post:
      properties:
        music_file:
          format: binary
          title: Music File
          type: string
      required:
      - music_file
      title: Body_predict_by_music_file_api_v1_predict_mp3_post
      type: object
    Body_upload_model_api_v1_text_upload_model_post:
      properties:
        model:
          format: binary
          title: Model
          type: string
      required:
      - model
      title: Body_upload_model_api_v1_text_upload_model_post
      type: object
    DatasetNameRequest:
      properties:
        name:
          title: Name
          type: string
      required:
      - name
      title: DatasetNameRequest
      type: object
    DatasetNameResponse:
      properties:
        message:
          title: Message
          type: string
      required:
      - message
      title: DatasetNameResponse
      type: object
    DatasetNamesResponse:
      properties:
        names:
          items:
            type: string
          title: Names
          type: array
      required:
      - names
      title: DatasetNamesResponse
      type: object
    FitRequest:
      properties:
        dataset_name:
          title: Dataset Name
          type: string
        epochs:
          title: Epochs
          type: integer
        learning_rate:
          title: Learning Rate
          type: number
      required:
      - epochs
      - learning_rate
      - dataset_name
      title: FitRequest
      type: object
    FitResponse:
      properties:
        model_number_id:
          title: Model Number Id
          type: string
        training_loss_history:
          items:
            type: number
          title: Training Loss History
          type: array
        y_pred:
          items:
            items:
              type: integer
            type: array
          title: Y Pred
          type: array
        y_true:
          items:
            items:
              type: integer
            type: array
          title: Y True
          type: array
      required:
      - y_true
      - y_pred
      - training_loss_history
      - model_number_id
      title: FitResponse
      type: object
    GenreModel:
      properties:
        count:
          title: Count
          type: integer
        genre:
          title: Genre
          type: string
        songs:
          items:
            type: string
          title: Songs
          type: array
      required:
      - genre
      - count
      - songs
      title: GenreModel
      type: object
    GenresResponse:
      properties:
        genres:
          items:
            type: string
          title: Genres
          type: array
      required:
      - genres
      title: GenresResponse
      type: object
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
    LabelsResponse:
      properties:
        labels:
          items:
            type: string
          title: Labels
          type: array
      required:
      - labels
      title: LabelsResponse
      type: object
    ModelNameRequest:
      properties:
        id:
          title: Id
          type: string
        name:
          title: Name
          type: string
      required:
      - name
      - id
      title: ModelNameRequest
      type: object
    ModelsNamesResponse:
      properties:
        names:
          items:
            type: string
          title: Names
          type: array
      required:
      - names
      title: ModelsNamesResponse
      type: object
    PredictByModelResponse:
      properties:
        genres:
          items:
            type: string
          title: Genres
          type: array
      required:
      - genres
      title: PredictByModelResponse
      type: object
    TextFromMP3:
      properties:
        language:
          title: Language
          type: string
        status:
          title: Status
          type: string
        text:
          title: Text
          type: string
      required:
      - text
      - language
      - status
      title: TextFromMP3
      type: object
    TopGenresResponse:
      properties:
        top_genres:
          items:
            $ref: '#/components/schemas/GenreModel'
          title: Top Genres
          type: array
      required:
      - top_genres
      title: TopGenresResponse
      type: object
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
info:
  title: FastAPI
  version: 0.1.0
openapi: 3.1.0
paths:
  /:
    get:
      operationId: docs_redirect__get
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
      summary: Docs Redirect
  /api/v1/clear_cache:
    post:
      operationId: clear_cache_api_v1_clear_cache_post
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '404':
          description: Not found
      summary: Clear Cache
      tags:
      - Music
  /api/v1/fit_model:
    post:
      operationId: fit_model_api_v1_fit_model_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FitRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FitResponse'
          description: Successful Response
        '404':
          description: Not found
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Fit Model
      tags:
      - Music
  /api/v1/get_datasets_names:
    get:
      operationId: get_datasets_names_api_v1_get_datasets_names_get
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatasetNamesResponse'
          description: Successful Response
        '404':
          description: Not found
      summary: Get Datasets Names
      tags:
      - Music
  /api/v1/get_labels:
    post:
      operationId: get_labels_api_v1_get_labels_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DatasetNameRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LabelsResponse'
          description: Successful Response
        '404':
          description: Not found
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Labels
      tags:
      - Music
  /api/v1/models_names:
    get:
      operationId: get_models_names_api_v1_models_names_get
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelsNamesResponse'
          description: Successful Response
        '404':
          description: Not found
      summary: Get Models Names
      tags:
      - Music
  /api/v1/predict:
    post:
      operationId: predict_api_v1_predict_post
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_predict_api_v1_predict_post'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PredictByModelResponse'
          description: Successful Response
        '404':
          description: Not found
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Predict
      tags:
      - Music
  /api/v1/predict_mp3:
    post:
      operationId: predict_by_music_file_api_v1_predict_mp3_post
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_predict_by_music_file_api_v1_predict_mp3_post'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenresResponse'
          description: Successful Response
        '404':
          description: Not found
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Predict By Music File
      tags:
      - Music
  /api/v1/save_model_name:
    post:
      operationId: save_model_name_api_v1_save_model_name_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModelNameRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatasetNameResponse'
          description: Successful Response
        '404':
          description: Not found
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Save Model Name
      tags:
      - Music
  /api/v1/text/get_text:
    post:
      operationId: get_text_api_v1_text_get_text_post
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_get_text_api_v1_text_get_text_post'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TextFromMP3'
          description: Successful Response
        '404':
          description: Not found
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Text
      tags:
      - Text
  /api/v1/text/predict:
    post:
      operationId: predict_api_v1_text_predict_post
      parameters:
      - in: query
        name: text
        required: true
        schema:
          title: Text
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PredictByModelResponse'
          description: Successful Response
        '404':
          description: Not found
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Predict
      tags:
      - Text
  /api/v1/text/upload_dataset:
    post:
      operationId: convert_files_to_dataframe_api_v1_text_upload_dataset_post
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_convert_files_to_dataframe_api_v1_text_upload_dataset_post'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '404':
          description: Not found
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Convert Files To Dataframe
      tags:
      - Text
  /api/v1/text/upload_model:
    post:
      operationId: upload_model_api_v1_text_upload_model_post
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_upload_model_api_v1_text_upload_model_post'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                title: Response Upload Model Api V1 Text Upload Model Post
                type: object
          description: Successful Response
        '404':
          description: Not found
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Upload Model
      tags:
      - Text
  /api/v1/top_genres:
    get:
      operationId: top_genres_api_v1_top_genres_get
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TopGenresResponse'
          description: Successful Response
        '404':
          description: Not found
      summary: Top Genres
      tags:
      - Music
  /api/v1/upload_dataset:
    post:
      operationId: convert_files_to_dataframe_api_v1_upload_dataset_post
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_convert_files_to_dataframe_api_v1_upload_dataset_post'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '404':
          description: Not found
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Convert Files To Dataframe
      tags:
      - Music